Fixed everything; stats actually work now

+++ backend/Onied/Courses/Courses.Data/Abstractions/IStatsRepository.cs
@@ -3,7 +3,7 @@ namespace Courses.Data.Abstractions;
 public interface IStatsRepository
 {
     Task CreateTablesIfNotExistsAsync();
-    Task<int> GetCourseLikesAsync(int courseId);
+    Task<ulong> GetCourseLikesAsync(int courseId);
     Task LikeCourseAsync(int courseId, Guid userId);
     Task UnlikeCourseAsync(int courseId, Guid userId);
     Task<bool> IsCourseLikedAsync(int courseId, Guid userId);
